1618372: Print accessible content paths from X509 extension using rct
The actual accessible content paths for an entitlement are encoded in a
binary extension within the entitlement certificate.  Previously, the
rct tool was not printing this information and instead relied on the
paths in the gzipped JSON blob that is adjacent to the actual
certificate in the PEM file.  In rare cases, the paths in the JSON and
the paths in the extension can be mismatched.  Debugging these cases is
exceptionally difficult since there is no easy way to decode the binary
extension.

This patch creates a method that can render the extension into a list of
human-readable paths and then adds a section to the rct output for
entitlement certs with these paths.